+++ core/data.lua
@@ -1,0 +1,189 @@
+local _, SRT = ...
+
+SRT.data = SRT.data or {}
+
+SRT.data.raceTimes = {}
+
+local raceTimes = SRT.data.raceTimes
+
+local questList = {
+    80219,
+    80220,
+    80221,
+    80222,
+    80223,
+    80224,
+    80225,
+    80226,
+    80227,
+    80228,
+    80229,
+    80230,
+    80231,
+    80232,
+    80233,
+    80234,
+    80235,
+    80236,
+    80237,
+    80238,
+    80239,
+    80240,
+    80242,
+    80243,
+    80244,
+    80245,
+    80246,
+    80247,
+    80248,
+    80249,
+    80250,
+    80251,
+    80252,
+    80253,
+    80254,
+    80255,
+    80256,
+    80257,
+    80258,
+    80259,
+    80260,
+    80261,
+    80265,
+    80266,
+    80267,
+    80268,
+    80269,
+    80270,
+    80271,
+    80272,
+    80273,
+    80274,
+    80275,
+    80276,
+    80277,
+    80278,
+    80279,
+    80280,
+    80281,
+    80282,
+    80283,
+    80284,
+    80285,
+    80286,
+    80287,
+    80288,
+    80289,
+    80290,
+    80291,
+    80292,
+    80293,
+    80294
+}
+
+SRT.data.questList = questList
+
+-- The War Within
+-- Isle of Dorn
+-- Dornogal Drift
+raceTimes[80219] = {48, 53}
+raceTimes[80225] = {43, 46}
+raceTimes[80231] = {43, 46}
+-- Storm Watch's Survey
+raceTimes[80220] = {63, 68}
+raceTimes[80226] = {60, 63}
+raceTimes[80232] = {62, 65}
+-- Basin Bypass
+raceTimes[80221] = {58, 63}
+raceTimes[80227] = {54, 57}
+raceTimes[80233] = {57, 60}
+-- The Wold Ways
+raceTimes[80222] = {68, 73}
+raceTimes[80228] = {68, 71}
+raceTimes[80234] = {70, 73}
+-- Thunderhead Trail
+raceTimes[80223] = {70, 75}
+raceTimes[80229] = {66, 69}
+raceTimes[80235] = {66, 69}
+-- Orecreg's Doglegs
+raceTimes[80224] = {65, 70}
+raceTimes[80230] = {61, 64}
+raceTimes[80236] = {61, 64}
+
+-- The Ringing Deeps
+-- Earthenworks Weave
+raceTimes[80237] = {52, 57}
+raceTimes[80244] = {49, 52}
+raceTimes[80250] = {50, 53}
+-- Ringing Deeps Ramble
+raceTimes[80238] = {57, 62}
+raceTimes[80245] = {53, 56}
+raceTimes[80251] = {53, 56}
+-- Chittering Concourse
+raceTimes[80239] = {56, 61}
+raceTimes[80246] = {53, 56}
+raceTimes[80252] = {54, 57}
+-- Cataract River Cruise
+raceTimes[80240] = {60, 65}
+raceTimes[80247] = {58, 61}
+raceTimes[80253] = {57, 60}
+-- Taelloch Twist
+raceTimes[80242] = {47, 52}
+raceTimes[80248] = {43, 46}
+raceTimes[80254] = {44, 47}
+-- Opportunity Point Amble
+raceTimes[80243] = {77, 82}
+raceTimes[80249] = {71, 74}
+raceTimes[80255] = {72, 75}
+
+-- Hallowfall
+-- Dunelle's Detour
+raceTimes[80256] = {65, 70}
+raceTimes[80265] = {62, 65}
+raceTimes[80271] = {64, 67}
+-- Tenir's Traversal
+raceTimes[80257] = {65, 70}
+raceTimes[80266] = {60, 63}
+raceTimes[80272] = {63, 66}
+-- Light's Redoubt Descent
+raceTimes[80258] = {63, 68}
+raceTimes[80267] = {62, 65}
+raceTimes[80273] = {62, 65}
+-- Stillstone Slalom
+raceTimes[80259] = {56, 61}
+raceTimes[80268] = {54, 57}
+raceTimes[80274] = {56, 59}
+-- Mereldar Meander
+raceTimes[80260] = {76, 81}
+raceTimes[80269] = {71, 74}
+raceTimes[80275] = {71, 74}
+-- Velhan's Venture
+raceTimes[80261] = {55, 60}
+raceTimes[80270] = {50, 53}
+raceTimes[80276] = {50, 53}
+
+-- Azj-Kahet
+-- City of Threads Twist
+raceTimes[80277] = {78, 83}
+raceTimes[80283] = {74, 77}
+raceTimes[80289] = {74, 77}
+-- Maddening Deep Dip
+raceTimes[80278] = {58, 63}
+raceTimes[80284] = {54, 57}
+raceTimes[80290] = {56, 59}
+-- The Weaver's Wing
+raceTimes[80279] = {54, 59}
+raceTimes[80285] = {51, 54}
+raceTimes[80291] = {50, 53}
+-- Rak-Ahat Rush
+raceTimes[80280] = {70, 75}
+raceTimes[80286] = {66, 69}
+raceTimes[80292] = {66, 69}
+-- Pit Plunge
+raceTimes[80281] = {63, 68}
+raceTimes[80287] = {61, 64}
+raceTimes[80293] = {61, 64}
+-- Siegehold Scuttle
+raceTimes[80282] = {70, 75}
+raceTimes[80288] = {66, 69}
+raceTimes[80294] = {63, 66}