+++ locals/localization.deDE.lua
@@ -6,3 +6,7 @@ local L = SRT.localization
 
 L["time"] = "Zeit"
 L["seconds"] = "Sekunden"
+L["gold-time"] = "|T616373:0|t Goldzeit: %s Sek."
+L["silver-time"] = "|T616375:0|t Silberzeit: %s Sek."
+L["bronze-time"] = "|T616372:0|t Bronzezeit"
+L["not-time"] = "keine Zeit verfügbar"