+++ locals/localization.enUS.lua
@@ -9,3 +9,7 @@ local L = SRT.localization
 
 L["time"] = "Time"
 L["seconds"] = "Seconds"
+L["gold-time"] = "|T616373:0|t Gold time: %s sec"
+L["silver-time"] = "|T616375:0|t Silver time: %s sec"
+L["bronze-time"] = "|T616372:0|t Bronze time"
+L["not-time"] = "no time available"