+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Skyriding Race Timer
-### [1.1.0](https://github.com/diomsg-code/SkyridingRaceTimer/tree/1.1.0) (2024-10-20)
-[Full Changelog](https://github.com/diomsg-code/SkyridingRaceTimer/compare/1.0.0...1.1.0) - [Previous Releases](https://github.com/diomsg-code/SkyridingRaceTimer/releases)
+### [1.2.0](https://github.com/diomsg-code/SkyridingRaceTimer/tree/1.2.0) (2024-10-21)
+[Full Changelog](https://github.com/diomsg-code/SkyridingRaceTimer/compare/1.1.0...1.2.0) - [Previous Releases](https://github.com/diomsg-code/SkyridingRaceTimer/releases)
 
-- Added: Language support
+- Added: If available, the gold and silver times for the current race are displayed
+- Added: Race times for the expansion 'The War Within'